+++ energyquantified/data/ohlc.py
@@ -6,6 +6,7 @@ import enum
 from .base import Series
 from ..metadata import ContractPeriod, OHLCField
 from ..time import Frequency
+from ..utils.pandas import ohlc_list_to_dataframe
 
 
 
@@ -81,7 +82,7 @@ class OHLC:
 
 class OHLCList(list):
     """
-    A collection of OHLC data . Can contain all sorts of contracts
+    A collection of OHLC data. Can contain all sorts of contracts
     (yearly, monthly, weekly etc.) for a specific market.
     """
 
@@ -91,8 +92,66 @@ class OHLCList(list):
         #: The curve holding these OHLC objects
         self.curve = curve
 
-    def __str__(self):
-        return f"<OHLCData: curve=\"{self.curve}\", items={len(self)}>"
+    def append(self, value):
+        raise NotImplementedError("OHLCList does not support append")
 
-    def has_data(self):
-        return bool(self.data)
+    def extend(self, iterable):
+        raise NotImplementedError("OHLCList does not support extend")
+
+    def insert(self, index, value):
+        raise NotImplementedError("OHLCList does not support insert")
+
+    def remove(self, value):
+        raise NotImplementedError("OHLCList does not support remove")
+
+    def pop(self, index=-1):
+        raise NotImplementedError("OHLCList does not support pop")
+
+    def clear(self):
+        raise NotImplementedError("OHLCList does not support clear")
+
+    def __setitem__(self, key, value):
+        raise NotImplementedError("OHLCList does not support __setitem__")
+
+    def __delitem__(self, key):
+        raise NotImplementedError("OHLCList does not support __delitem__")
+
+    def __add__(self, rhs):
+        raise NotImplementedError("OHLCList does not support __add__")
+
+    def __iadd__(self, rhs):
+        raise NotImplementedError("OHLCList does not support __iadd__")
+
+    def __mul__(self, rhs):
+        raise NotImplementedError("OHLCList does not support multiply")
+
+    def __rmul__(self, rhs):
+        raise NotImplementedError("OHLCList does not support multiply")
+
+    def __imul__(self, rhs):
+        raise NotImplementedError("OHLCList does not support multiply")
+
+    def __str__(self):
+        return f"<OHLCList: curve=\"{self.curve}\", items={super().__str__()}>"
+
+    def to_df(self):
+        """
+        Convert this OHLCList to a ``pandas.DataFrame`` with a column for
+        open, high, low, close, settlement, volume and open interest.
+
+        :return: A DataFrame
+        :rtype: pandas.DataFrame
+        :raises ImportError: When pandas is not installed on the system
+        """
+        return self.to_dataframe()
+
+    def to_dataframe(self):
+        """
+        Convert this OHLCList to a ``pandas.DataFrame`` with a column for
+        open, high, low, close, settlement, volume and open interest.
+
+        :return: A DataFrame
+        :rtype: pandas.DataFrame
+        :raises ImportError: When pandas is not installed on the system
+        """
+        return ohlc_list_to_dataframe(self)